added TextureLibrary and other changes

+++ CrudeRender/src/Crude/Core/Timestep.h
@@ -14,17 +14,17 @@ namespace Crude
     class Timestep
     {
     public:
-        Timestep(float time = 0.0f)
+        Timestep(double time = 0.0f)
         : m_Time(time)
         {
         }
         
-        operator float() const { return m_Time; }
+        operator double() const { return m_Time; }
         
         float GetSeconds() const { return m_Time; }
-        float GetMilliseconds() const { return m_Time * 1000.0f; }
+        float GetMilliseconds() const { return m_Time * 1000.0; }
     private:
-        float m_Time;
+        double m_Time;
     };
 }
